# Conflicts:
#	CMakeLists.txt
#	Data/Sys/GameSettings/G8ME01.ini
#	Data/Sys/GameSettings/NAC.ini
#	Data/Sys/GameSettings/NAR.ini
#	Data/Sys/GameSettings/RM3P01.ini
#	Data/Sys/totaldb.dsy
#	Source/Core/AudioCommon/WaveFile.cpp
#	Source/Core/Common/Common.h
#	Source/Core/Core/ActionReplay.cpp
#	Source/Core/Core/Boot/Boot.cpp
#	Source/Core/Core/Boot/Boot_BS2Emu.cpp
#	Source/Core/Core/Boot/Boot_WiiWAD.cpp
#	Source/Core/Core/BootManager.cpp
#	Source/Core/Core/ConfigManager.cpp
#	Source/Core/Core/ConfigManager.h
#	Source/Core/Core/Core.cpp
#	Source/Core/Core/Core.h
#	Source/Core/Core/Core.vcxproj
#	Source/Core/Core/Core.vcxproj.filters
#	Source/Core/Core/HW/DSP.h
#	Source/Core/Core/HW/DSPHLE/DSPHLE.cpp
#	Source/Core/Core/HW/DSPHLE/UCodes/UCodes.h
#	Source/Core/Core/HW/DSPHLE/UCodes/Zelda.cpp
#	Source/Core/Core/HW/GCPadEmu.cpp
#	Source/Core/Core/HW/SystemTimers.cpp
#	Source/Core/Core/HW/VideoInterface.cpp
#	Source/Core/Core/HW/WiimoteEmu/Attachment/Attachment.h
#	Source/Core/Core/HW/WiimoteEmu/Attachment/Classic.cpp
#	Source/Core/Core/HW/WiimoteEmu/Attachment/Classic.h
#	Source/Core/Core/HW/WiimoteEmu/Attachment/Nunchuk.cpp
#	Source/Core/Core/HW/WiimoteEmu/Attachment/Nunchuk.h
#	Source/Core/Core/HW/WiimoteEmu/WiimoteEmu.cpp
#	Source/Core/Core/HW/WiimoteEmu/WiimoteEmu.h
#	Source/Core/Core/IPC_HLE/WII_IPC_HLE.cpp
#	Source/Core/Core/IPC_HLE/WII_IPC_HLE_Device_es.cpp
#	Source/Core/Core/IPC_HLE/WII_IPC_HLE_Device_net_ssl.cpp
#	Source/Core/Core/IPC_HLE/WII_IPC_HLE_Device_usb_bt_base.cpp
#	Source/Core/Core/IPC_HLE/WII_IPC_HLE_Device_usb_bt_base.h
#	Source/Core/Core/IPC_HLE/WII_IPC_HLE_Device_usb_bt_real.h
#	Source/Core/Core/PowerPC/Interpreter/Interpreter_Branch.cpp
#	Source/Core/Core/PowerPC/Jit64Common/FarCodeCache.h
#	Source/Core/Core/PowerPC/Jit64Common/TrampolineCache.h
#	Source/Core/Core/PowerPC/JitCommon/JitCache.h
#	Source/Core/Core/PowerPC/PPCAnalyst.cpp
#	Source/Core/DolphinQt2/DolphinQt2.vcxproj
#	Source/Core/DolphinQt2/GameList/GameFile.h
#	Source/Core/DolphinWX/DolphinWX.vcxproj
#	Source/Core/DolphinWX/DolphinWX.vcxproj.filters
#	Source/Core/DolphinWX/Frame.cpp
#	Source/Core/DolphinWX/ISOFile.cpp
#	Source/Core/DolphinWX/Input/HotkeyInputConfigDiag.cpp
#	Source/Core/DolphinWX/Input/InputConfigDiag.cpp
#	Source/Core/InputCommon/CMakeLists.txt
#	Source/Core/InputCommon/ControllerEmu.cpp
#	Source/Core/InputCommon/ControllerEmu.h
#	Source/Core/InputCommon/ControllerInterface/ControllerInterface.cpp
#	Source/Core/InputCommon/ControllerInterface/Device.h
#	Source/Core/InputCommon/InputCommon.vcxproj
#	Source/Core/InputCommon/InputCommon.vcxproj.filters
#	Source/Core/VideoBackends/D3D/FramebufferManager.cpp
#	Source/Core/VideoBackends/D3D/FramebufferManager.h
#	Source/Core/VideoBackends/D3D/Render.cpp
#	Source/Core/VideoBackends/D3D12/Render.cpp
#	Source/Core/VideoBackends/OGL/Render.cpp
#	Source/Core/VideoBackends/OGL/VertexManager.h
#	Source/Core/VideoBackends/Software/SWRenderer.cpp
#	Source/Core/VideoCommon/BPStructs.cpp
#	Source/Core/VideoCommon/OpcodeDecoding.cpp
#	Source/Core/VideoCommon/RenderBase.cpp
#	Source/Core/VideoCommon/TextureCacheBase.cpp
#	Source/Core/VideoCommon/TextureCacheBase.h
#	Source/Core/VideoCommon/TextureDecoder_x64.cpp
#	Source/Core/VideoCommon/VertexShaderManager.cpp
#	Source/Core/VideoCommon/VideoCommon.vcxproj
#	Source/Core/VideoCommon/VideoCommon.vcxproj.filters
#	Source/Core/VideoCommon/VideoConfig.cpp
#	Source/UnitTests/UnitTests.vcxproj

+++ Source/Core/Core/PowerPC/Jit64Common/FarCodeCache.cpp
@@ -4,7 +4,7 @@
 
 #include "Core/PowerPC/Jit64Common/FarCodeCache.h"
 
-void FarCodeCache::Init(int size)
+void FarCodeCache::Init(size_t size)
 {
   AllocCodeSpace(size);
   m_enabled = true;
