+++ Source/Core/Core/PowerPC/Jit64Common/TrampolineCache.h
@@ -4,6 +4,7 @@
 
 #pragma once
 
+#include <cstddef>
 #include "Common/CommonTypes.h"
 #include "Core/PowerPC/Jit64Common/EmuCodeBlock.h"
 
@@ -17,8 +18,8 @@ struct TrampolineInfo;
 #endif
 // a bit of a hack; the MMU results in more code ending up in the trampoline cache,
 // because fastmem results in far more backpatches in MMU mode
-CONSTEXPR(int, TRAMPOLINE_CODE_SIZE, 1024 * 1024 * 8);
-CONSTEXPR(int, TRAMPOLINE_CODE_SIZE_MMU, 1024 * 1024 * 32);
+CONSTEXPR(size_t, TRAMPOLINE_CODE_SIZE, 1024 * 1024 * 8);
+CONSTEXPR(size_t, TRAMPOLINE_CODE_SIZE_MMU, 1024 * 1024 * 32);
 
 // We need at least this many bytes for backpatching.
 CONSTEXPR(int, BACKPATCH_SIZE, 5);
@@ -29,7 +30,7 @@ class TrampolineCache : public EmuCodeBlock
   const u8* GenerateWriteTrampoline(const TrampolineInfo& info);
 
 public:
-  void Init(int size);
+  void Init(size_t size);
   void Shutdown();
   const u8* GenerateTrampoline(const TrampolineInfo& info);
   void ClearCodeSpace();
